requirements.txt
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
faunadb==4.3.1
flake8==6.0.0
future==0.18.2
h2==2.6.2
hpack==3.0.0
hyper==0.7.0
hyperframe==3.2.0
idna==3.4
iso8601==1.1.0
mccabe==0.7.0
openskill==4.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
python-dotenv==0.21.0
requests==2.28.1
tomli==2.0.1
tqdm==4.64.1
urllib3==1.26.13
